Fix the island manager to take joints into account.
The island manager ignored completely joints when building its islands
and waking up bodies.  This is fixed: activating a rigid body will
activate everything attachet to itself by a joint.

As a side effect, this adds a trat `Joint` that exposes methods to get
the joint anchor. This is temporary, until a true `Joint` trait is
implemented to abstract everything about joint.

Adds a 3d ragdoll demo for testing.