Temporary fix for the 2d ball-in-socket joint.
Since nalgebra does not have staticly-sized rectangular matrices yet, the `col` operation required
by the ball-in-socket joint torque axis is not available in 2d.

This uses the `row` operation instead. To avoid breaking the 3d case, signs are inverted (the
matrix from which we take the column is scew-symetric so: row(i) == -col(i)).