implements send and sync for world
(examples and test have not been updated)

The whole point was not to use Rc<RefCell<_>> for rigid bodies, sensor
and joint and instead use a storage and an index to handle them.

The storage must be efficient index_vector is my implementation:
it uses a vector to store the data and never move the position of the
elements in the vector. So the index of the vector always point to the
same element.
We use an offset in order to make sensor and rigid_bodies index
different.

still issue:
* fitler data is difficult to use as we can't send a tuple of
  reference to storage because it would imply a use of lifetime that the
  compiler can't handle (or it's me). I wanted `FilterData = (&'a
  RigidBodyStorage, &'a SensorStorage) but I couldn't impl it.

  I think the filter things could be different like instead of register
  callbacks in collision world. callbacks would be stored in world and
  collision_world.perform_broad_phase would have a unique callback
  argument. By doing so all the reference lifetimes would be in the in the
  function and so it should be OK.

  the only drawback of this is that it is to the world to ensure to call
  `broad_phase.deferred_recompute_all_proximities();`

still todo:
* documentation
* maybe wrap sensor, joint and rigid_body index in a struct(usize) so user
  can't missuse them.
* API