Utilize BodySet as a trait object instead of param
This commit prevents BodySet from polluting as a type parameter into all
the core types required for physics stepping. Thanks to this, problems
involved in dealing with these types in Specs usage are alleviated.
However, this commit also removes the associated type of Body and always
assumes dynamic dispatch on the Body type.